Back-End/requirements.txt
djangorestframework==3.15.2
django==5.0.7
djangorestframework-simplejwt==5.3.1 